long last commit, working as hell on editor. Never gonna finish in time :<, but i'll try

+++ TestGame/Include/Jamgame.h
@@ -35,8 +35,8 @@ private: // Variable
 	// Other
 	GameTimer*	m_gameTimer;
 	bool		m_gamePaused;
-	int			m_mousePositionX;
-	int			m_mousePositionY;
+	float		m_mousePositionX;
+	float		m_mousePositionY;
 	bool		m_mouseClicked;
 
 	// Game
