+++ TestGame/Include/Scene/BaseScene.h
@@ -13,7 +13,7 @@ public:
 	virtual ~BaseScene();	
 
 	virtual void Initialize(SceneManagerInterface* p_sceneManagerInteface, Jamgine::JamgineEngine* p_engine);
-	virtual void Update(double p_deltaTime, int p_mousePositionX, int p_mousePositionY, bool p_lMouseClicked) = 0;
+	virtual void Update(double p_deltaTime, float p_mousePositionX, float p_mousePositionY, bool p_lMouseClicked) = 0;
 	virtual void Render() = 0;
 
 protected:
