+++ TestGame/Include/Scene/MainMenuScene.h
@@ -10,6 +10,6 @@ public:
 
 	int i;
 	virtual void Initialize(SceneManagerInterface* p_sceneManagerInteface,Jamgine::JamgineEngine* p_engine);
-	void Update(double p_deltaTime, int p_mousePositionX, int p_mousePositionY, bool p_lMouseClicked);
+	void Update(double p_deltaTime, float p_mousePositionX, float p_mousePositionY, bool p_lMouseClicked);
 	void Render();
 };
